requirements.txt
Django>=4.2,<5.0
scikit-learn>=1.3.0
numpy>=1.24.0
python-docx>=1.1.0
PyPDF2>=3.0.0
spacy>=3.7.0
nltk>=3.8.1
Pillow>=10.0.0
